+++ fe_flutter/lib/model/user_model.dart
@@ -1,3 +1,4 @@
+
 class User {
   String? secretEmail;
   String? secretPassword;
