+++ fe_flutter/lib/screens/bottomMenu.dart
@@ -1,6 +1,6 @@
-import 'package:fe_flutter/mainPage.dart';
-import 'package:fe_flutter/myPage.dart';
-import 'package:fe_flutter/myRefrigerator.dart';
+import 'package:fe_flutter/screens/mainPage.dart';
+import 'package:fe_flutter/screens/myPage.dart';
+import 'package:fe_flutter/screens/myRefrigerator.dart';
 import 'package:flutter/material.dart';
 
 class BottomMenu extends StatefulWidget {
