+++ fe_flutter/lib/screens/login/join.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:fe_flutter/login/buildTextFormField.dart';
+import 'package:fe_flutter/screens/login/buildTextFormField.dart';
 
 class JoinPage extends StatefulWidget {
   @override
