+++ fe_flutter/lib/screens/myRefrigerator.dart
@@ -1,4 +1,4 @@
-import 'package:fe_flutter/ingredientMoreInfo/ingredientMoreInfo.dart';
+import 'package:fe_flutter/screens/ingredientMoreInfo/ingredientMoreInfo.dart';
 import 'package:flutter/material.dart';
 // 임시 페이지입니다.
 class MyRefrigerator extends StatelessWidget {
