+++ fe_flutter/lib/screens/register/allergy.dart
