+++ fe_flutter/lib/screens/register/category.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:fe_flutter/register/customIconButton.dart';
+import 'package:fe_flutter/screens/register/customIconButton.dart';
 
 class CategoryPage extends StatefulWidget {
   @override
